slv-adf-functions/requirements.txt
-i https://pypi.org/simple
azure-common==1.1.28
azure-core==1.26.2 ; python_version >= '3.7'
azure-functions==1.12.0
azure-functions-worker==1.1.9
azure-identity==1.12.0
azure-keyvault-secrets==4.6.0
certifi==2022.12.7 ; python_version >= '3.6'
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.0 ; python_version >= '3.6'
grpcio==1.33.2
grpcio-tools==1.33.2
idna==3.4 ; python_version >= '3.5'
isodate==0.6.1
msal==1.20.0
msal-extensions==1.0.0
msrest==0.7.1 ; python_version >= '3.6'
oauthlib==3.2.2 ; python_version >= '3.6'
portalocker==2.7.0 ; python_version >= '3.5' and platform_system == 'Windows'
protobuf==3.20.3 ; python_version >= '3.7'
psycopg2==2.9.5
pycparser==2.21
pyjwt[crypto]==2.6.0 ; python_version >= '3.7'
pyodbc==4.0.35
python-dotenv==0.21.1
pywin32==305 ; platform_system == 'Windows'
requests==2.28.2
requests-oauthlib==1.3.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
typing-extensions==4.4.0 ; python_version >= '3.7'
urllib3==1.26.14 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
